// src/app/utopia/engines/enums/sprite-type.ts
export enum UtopiaSpriteType {
  Player1 = 'Player1',
  Player2 = 'Player2',
  Player3 = 'Player3',
  Player1i = 'Player1i',
  Player2i = 'Player2i',
  Player3i = 'Player3i',
  Tiles = 'Tiles',
  Items = 'Items',
  Hud = 'Hud',
  Enemies = 'Enemies',
  Enemiesi = 'Enemiesi',
  Moon = 'Moon ',
  MoonShiny = 'MoonShiny',
  Sun = 'Sun'
}